requirements.txt
asgiref==3.4.1
astroid==2.8.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.4
coverage==7.3.2
cryptography==41.0.4
dill==0.3.5.1
Django==3.2.7
django-admin-rangefilter==0.10.0
django-adminlte-ui==1.5.1
django-ajax-selects==2.1.0
django-constance==3.1.0
django-cors-headers==3.14.0
django-currentuser==0.5.3
django-environ==0.10.0
django-filter==22.1
django-jazzmin==2.6.0
django-picklefield==3.1
django-reversion==5.0.4
django-treebeard==4.3
django-utils-six==2.0
django-webtest==1.9.11
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.4.1
humanize==4.6.0
idna==3.4
importlib-metadata==6.1.0
isort==5.9.3
itsdangerous==0.24
josepy==1.13.0
lazy-object-proxy==1.6.0
Markdown==3.4.3
mccabe==0.6.1
mozilla-django-oidc==3.0.0
openrouteservice==2.3.3
Pillow==9.3.0
platformdirs==2.4.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodomex==3.19.0
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.11.1
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==23.2.0
pytz==2021.1
redis==3.2.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typed-ast==1.5.4
typing-extensions==3.10.0.2
urllib3==2.0.5
waitress==2.1.2
WebOb==1.8.7
webservices==0.7
WebTest==3.0.0
wrapt==1.12.1
zipp==3.15.0
uWSGI==2.0.23